renaming of tcp channel class/filename
